Complete overhaul of analyzer.py
Previous file was a general blueprint. This update is capable of taking user input of date (year, day, month), scan every .txt file that has the date in the name "20YY-DD-MM_H" (where H ranges from 0 to 23), and generate/save two plots that compare temperature and humidity values from each sensor.

Note: program throws a generic error if given date does not match .txt files and if there isn't a complete set of 24 .txt files. Working on separating these errors, as well as finding a workaround for not having 24 .txt files of the same date with different hours.